fix: implement workaround for IntersectionObserver bug
Most browsers have a broken implementation of IntersectionObserver that does not recognise zero-area intersections as being isIntersecting. Additionally, they do not emit an entry when the element is already partially overlapping the viewport until the element is fully contained in the viewport. The lazy loader should be initialised after first render so that the element has some chance to be at a representative size and position.